// src/sdks/Newcomputation/computationV2ApiDocs.ts
export type getComputationAuditPageProps = {
  auditStatus?: '0' | '1' | '2' | '3' | '4' | null | undefined;
  pageNum: number;
  pageSize: number;
};
// ...
export interface AuditData {
  auditStatus?: '0' | '1' | '2' | '3' | '4';

  auditType?: '1' | '2';

  companyId?: number;

  createBy?: number;

  createByName?: string;

  createTime?: Date;

  dataId?: string;

  deleted?: boolean;

  id?: number;

  orgId?: number;

  orgName?: string;

  targetNames?: string;

  updateBy?: number;

  updateTime?: Date;

  userBtnFlag?: boolean;
  auditStatus_name?: string;
}